backend/dj_app/requirements.txt
Django
python-socketio
eventlet
redis
djangorestframework
drf-yasg
celery